python/agents/claim-adjudication-agent/pyproject.toml
[project]
name = "claim-adjudication-agent"
version = "1.0.0"
description = "AI-driven agent designed to facilitate health insurance claim processing and adjudication."
authors = [{ name = "Sandip Gupta", email = "sandipgupta@google.com" }, { name = "Ajitesh Kumar", email = "ajiteshk@google.com" }, { name = "Gaurav Jagavkar", email = "gauravcj@google.com" }]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.142.0",
    "google-genai>=1.68.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "google-adk>=1.27.0",
    "google-cloud-storage>=3.10.1",
    "google-cloud-logging>=3.14.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "google-adk[eval]>=1.27.0",
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.142.0",
    "nest-asyncio>=1.6.0",
    "pandas>=2.2.3,<3.0.0",
    "tabulate>=0.10.0",
    "agent-starter-pack>=0.14.1",
    "ruff>=0.15.7"
]
deployment = [
    "absl-py>=2.1.0",
]

[tool.ruff]
# extends the ruff rules established in the root .toml
extend = "../../../pyproject.toml"

[tool.ruff.lint]
# Ignores function complication rule throughout sample
ignore = ["C901", "PLR0915"]

[tool.ruff.lint.per-file-ignores]
# ignores unused import rule exclusively for this sample’s __init__.py files
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["app", "claim_adjudication_agent"]

[tool.hatch.build.targets.wheel]
packages = ["claim_adjudication_agent"]

[tool.hatch.build.targets.wheel.shared-data]
"claim_adjudication_agent/static" = "share/claim_adjudication_agent/static"
"claim_adjudication_agent/.env" = "share/claim_adjudication_agent/.env.template"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.agent-starter-pack]
example_question = "Adjudicate the claim ID CLAIMIDX0001"

[tool.agent-starter-pack.settings]
agent_directory = "claim_adjudication_agent"
